refactor: prepare to move dispatcher servlet to mvc package
