+++ app/src/main/java/com/techcourse/DispatcherServlet.java
@@ -1,13 +1,13 @@
 package com.techcourse;
 
+import com.interface21.webmvc.servlet.HandlerAdapter;
+import com.interface21.webmvc.servlet.HandlerExceptionResolver;
+import com.interface21.webmvc.servlet.HandlerMapping;
 import com.interface21.webmvc.servlet.ModelAndView;
 import com.interface21.webmvc.servlet.View;
-import com.interface21.webmvc.servlet.mvc.tobe.HandlerAdapter;
-import com.interface21.webmvc.servlet.mvc.tobe.HandlerAdapterRegistry;
-import com.interface21.webmvc.servlet.mvc.tobe.HandlerMapping;
-import com.interface21.webmvc.servlet.mvc.tobe.HandlerMappingRegistry;
-import com.interface21.webmvc.servlet.mvc.tobe.exception.ArgumentHandlerExceptionResolver;
-import com.interface21.webmvc.servlet.mvc.tobe.exception.HandlerExceptionResolver;
+import com.interface21.webmvc.servlet.handler.HandlerAdapterRegistry;
+import com.interface21.webmvc.servlet.handler.HandlerMappingRegistry;
+import com.interface21.webmvc.servlet.resolver.ArgumentHandlerExceptionResolver;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
